AIEngineering workflow toolkit/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-engineering-workflow-toolkit"
version = "1.0.0"
description = "Governed Code Review Infrastructure for Claude Code"
requires-python = ">=3.11"
dependencies = [
    # Claude API
    "anthropic>=0.40.0",
    # MCP server
    "fastmcp>=0.4.0",
    # OpenTelemetry
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    # Data validation
    "pydantic>=2.7.0",
    # CLI
    "click>=8.1.0",
    # Git integration
    "gitpython>=3.1.43",
    # Async HTTP
    "aiohttp>=3.9.0",
    # Deterministic tools (installed separately as CLI tools)
    # ruff, mypy, bandit — expected on PATH
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "bandit>=1.7.8",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
aiwt = "main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["eval*", "skills*", "subagents*", "mcp_server*", "orchestrator*", "review_agent*", "observability*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
